laboratory/requirements.txt
Django==4.2.6
django-dry-tests==0.2.1